@@ -14,13 +14,14 @@
     "@headlessui/react": "^1.7.4",
     "@headlessui/tailwindcss": "^0.1.1",
     "@hookform/resolvers": "^2.9.10",
+    "@proton/ual-webauth": "^0.1.18",
     "@types/react-beautiful-dnd": "^13.1.3",
     "@waxio/waxjs": "^1.1.0",
     "axios": "^1.1.3",
     "eosjs": "^22.1.0",
     "framer-motion": "^9.0.3",
     "is-ipfs": "^7.0.3",
-    "next": "13.0.3",
+    "next": "13.2.3",
     "phosphor-react": "^1.4.1",
     "react": "18.2.0",
     "react-beautiful-dnd": "^13.1.1",
