@@ -1,7 +1,8 @@
 import { Anchor } from 'ual-anchor';
 import { Wax } from '@eosdacio/ual-wax';
+import { WebAuth } from '@proton/ual-webauth';
 
-module.exports = {
+export default {
   eos: {
     name: 'EOS',
     imageUrl: 'https://bloks.io/img/chains/eos.png',
@@ -12,7 +13,6 @@ module.exports = {
     host: process.env.NEXT_PUBLIC_EOS_MAINNET_HOST,
     port: process.env.NEXT_PUBLIC_EOS_MAINNET_PORT,
   },
-
   wax: {
     name: 'WAX',
     imageUrl: 'https://wax.bloks.io/img/chains/wax.png',
@@ -23,7 +23,6 @@ module.exports = {
     host: process.env.NEXT_PUBLIC_WAX_MAINNET_HOST,
     port: process.env.NEXT_PUBLIC_WAX_MAINNET_PORT,
   },
-
   'wax-test': {
     name: 'WAX (Testnet)',
     imageUrl: 'https://wax.bloks.io/img/chains/wax.png',
@@ -34,7 +33,6 @@ module.exports = {
     host: process.env.NEXT_PUBLIC_WAX_TESTNET_HOST,
     port: process.env.NEXT_PUBLIC_WAX_TESTNET_PORT,
   },
-
   jungle4: {
     name: 'Jungle4 (EOS Testnet)',
     imageUrl: 'https://bloks.io/img/chains/jungle.png',
@@ -45,4 +43,24 @@ module.exports = {
     host: process.env.NEXT_PUBLIC_EOS_JUNGLE4_HOST,
     port: process.env.NEXT_PUBLIC_EOS_JUNGLE4_PORT,
   },
+  proton: {
+    name: 'Proton',
+    imageUrl: 'https://proton.bloks.io/img/chains/proton.png',
+    authenticators: [WebAuth, Anchor],
+    aaEndpoint: process.env.NEXT_PUBLIC_PROTON_MAINNET_AA_ENDPOINT,
+    chainId: process.env.NEXT_PUBLIC_PROTON_MAINNET_CHAIN_ID,
+    protocol: process.env.NEXT_PUBLIC_PROTON_MAINNET_PROTOCOL,
+    host: process.env.NEXT_PUBLIC_PROTON_MAINNET_HOST,
+    port: process.env.NEXT_PUBLIC_PROTON_MAINNET_PORT,
+  },
+  'proton-test': {
+    name: 'Proton (Testnet)',
+    imageUrl: 'https://proton.bloks.io/img/chains/proton.png',
+    authenticators: [WebAuth, Anchor],
+    aaEndpoint: process.env.NEXT_PUBLIC_PROTON_TESTNET_AA_ENDPOINT,
+    chainId: process.env.NEXT_PUBLIC_PROTON_TESTNET_CHAIN_ID,
+    protocol: process.env.NEXT_PUBLIC_PROTON_TESTNET_PROTOCOL,
+    host: process.env.NEXT_PUBLIC_PROTON_TESTNET_HOST,
+    port: process.env.NEXT_PUBLIC_PROTON_TESTNET_PORT,  
+  }
 };
