@@ -1,5 +1,5 @@
-import { appName } from '@configs/globalsConfig';
-import * as chainsConfig from '@configs/chainsConfig';
+import { appName, requestAccount } from '@configs/globalsConfig';
+import chainsConfig from '@configs/chainsConfig';
 import { blockchains } from '@utils/blockchains';
 
 export const authenticators = Object.keys(chainsConfig).reduce(
@@ -16,6 +16,12 @@ export const authenticators = Object.keys(chainsConfig).reduce(
           new Authenticator([blockchain], {
             appName,
             disableGreymassFuel: true,
+            transportOptions: {
+              requestAccount
+            },
+            selectorOptions: {
+              appName
+            }
           })
       ),
     };
